fixes compile problem
two lines above this one, you already checked if pSlash is NULL.  I feel pretty certain you're checking that you haven't gone past the string's end with this test, which means you intend to check the value of the byte at pSlash rather than the pointer.  This does address the compiler's complaint with this code.